+++ variables.tf
@@ -99,14 +99,10 @@ variable "existing_kms_instance_guid" {
   description = "The GUID of the Hyper Protect or Key Protect instance in which the key specified in `kms_key_crn` is coming from. Only required if `skip_iam_authorization_policy` is 'false'."
   type        = string
   default     = null
-  validation {
-    condition     = !var.skip_iam_authorization_policy && var.existing_kms_instance_guid == null ? false : true
-    error_message = "When var.skip_iam_authorization_policy is set to false, a value must be passed for var.existing_kms_instance_guid in order to create the auth policy."
-  }
 
   validation {
-    condition     = var.kms_encryption_enabled && var.existing_kms_instance_guid == null ? false : true
-    error_message = "When setting var.kms_encryption_enabled to true, a value must be passed for var.existing_kms_instance_guid"
+    condition     = var.kms_encryption_enabled && var.existing_kms_instance_guid && !var.skip_iam_authorization_policy == null ? false : true
+    error_message = "When var.skip_iam_authorization_policy is set to false, and var.kms_encryption_enabled to true, a value must be passed for var.existing_kms_instance_guid in order to create the auth policy."
   }
 }
